chore(aggregated-assets): move the domain layer behind re-export shims
Relocate the dada-client entity types and API layer into the aggregated-assets
DDD packages (LIVE-35226, epic LIVE-35223). Behaviour-neutral: zero consumer
files change, all 82 consumers keep working through one-line shims at the old
paths.

Pure relocation - no symbol renames, no signature changes, no zod, and no
endpoint restructuring. The per-use-case injectEndpoints split is LIVE-35301.

  entity-aggregated-asset  CryptoAssetMeta
  entity-interest-rate     InterestRate, ApyType
  api-aggregated-assets    wire schema, transforms, api, internals, errors, mocks

transforms.ts is split out of api.ts so the wire->entity boundary is separately
testable, matching the market-sentiment precedent.

NetworkInfo and CurrenciesOrder land in the api package rather than an entity: a
network IS a chain, already modelled by entity-currency-crypto, and
CurrenciesOrder is server sort metadata. Both were entities in the original
scoping; the naming review corrected that.

Breaks the last boundary violation. dadaIdToMarketId and the market item type
are copied from libs/ledger-live-common/src/market into internals/market.ts,
since a domain/* package cannot import legacy libs/*. Drift risk is documented
there and in the README: every field is optional, so divergence will never
produce a type error.

reducerPath stays the literal "assetsDataApi" and is now pinned by a test.
createCurrencyDataSelector hand-scans that key as a string and Storybook stories
preload it, so a rename would silently return undefined for every market and
interest-rate lookup with no type error.

The characterization tests from LIVE-35224 pass unmodified: 21 suites, 245
tests.
