+++ .changeset/quiet-comets-relocate.md
@@ -1,0 +1,8 @@
+---
+"@domain/entity-aggregated-asset": minor
+"@domain/entity-interest-rate": minor
+"@domain/api-aggregated-assets": minor
+"@ledgerhq/live-common": patch
+---
+
+Move the dada-client domain layer into the aggregated-assets DDD packages, leaving re-export shims at the old paths so no consumer changes
