+++ domain/api/aggregated-assets/src/stocks.mock.ts
@@ -1,0 +1,78 @@
+import { CryptoCurrencyIdSchema } from "@domain/entity-currency-crypto";
+import { TokenCurrencyIdSchema } from "@domain/entity-currency-token";
+
+export const mockStocksResponse = {
+  cryptoAssets: {
+    "urn:crypto:meta-currency:applex": {
+      id: "urn:crypto:meta-currency:applex",
+      ticker: "AAPLX",
+      name: "Apple xStock",
+      assetsIds: {
+        solana: "solana/spl/applex",
+      },
+    },
+    "urn:crypto:meta-currency:teslax": {
+      id: "urn:crypto:meta-currency:teslax",
+      ticker: "TSLAX",
+      name: "Tesla xStock",
+      assetsIds: {
+        solana: "solana/spl/teslax",
+      },
+    },
+  },
+  networks: {
+    solana: { id: "solana", name: "Solana" },
+  },
+  cryptoOrTokenCurrencies: {
+    "solana/spl/applex": {
+      type: "TokenCurrency" as const,
+      id: TokenCurrencyIdSchema.parse("solana/spl/applex"),
+      name: "Apple xStock",
+      ticker: "AAPLX",
+      contractAddress: "XsAAPL000000000000000000000000000000000000",
+      parentCurrencyId: CryptoCurrencyIdSchema.parse("solana"),
+      tokenType: "spl",
+      units: [
+        {
+          name: "AAPLX",
+          code: "AAPLX",
+          magnitude: 8,
+        },
+      ],
+    },
+    "solana/spl/teslax": {
+      type: "TokenCurrency" as const,
+      id: TokenCurrencyIdSchema.parse("solana/spl/teslax"),
+      name: "Tesla xStock",
+      ticker: "TSLAX",
+      contractAddress: "XsTSLA000000000000000000000000000000000000",
+      parentCurrencyId: CryptoCurrencyIdSchema.parse("solana"),
+      tokenType: "spl",
+      units: [
+        {
+          name: "TSLAX",
+          code: "TSLAX",
+          magnitude: 8,
+        },
+      ],
+    },
+  },
+  interestRates: {},
+  markets: {
+    "urn:crypto:meta-currency:applex": {
+      price: 226.4,
+      marketCap: 3_400_000_000_000,
+      priceChangePercentage24h: 1.2,
+    },
+    "urn:crypto:meta-currency:teslax": {
+      price: 248.5,
+      marketCap: 790_000_000_000,
+      priceChangePercentage24h: -0.8,
+    },
+  },
+  currenciesOrder: {
+    key: "marketCap",
+    order: "desc",
+    metaCurrencyIds: ["urn:crypto:meta-currency:applex", "urn:crypto:meta-currency:teslax"],
+  },
+};
