+++ domain/entity/aggregated-asset/src/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { CryptoAssetMeta } from "./schema";
