+++ domain/entity/interest-rate/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export type { InterestRate } from "./schema";
+export type { ApyType } from "./types";
