+++ domain/entity/interest-rate/src/types.ts
@@ -1,0 +1,2 @@
+/** The rate kinds the apps understand. Anything else is dropped at the consumer boundary. */
+export type ApyType = "NRR" | "APY" | "APR";
