+++ libs/ledger-live-common/package.json
@@ -263,14 +263,17 @@
   "dependencies": {
     "@blooo/hw-app-acre": "^1.1.1",
     "@cardano-foundation/ledgerjs-hw-app-cardano": "^7.1.2",
+    "@domain/api-aggregated-assets": "workspace:^",
     "@domain/api-currency-token": "workspace:^",
     "@domain/api-swap-quotes": "workspace:^",
+    "@domain/entity-aggregated-asset": "workspace:^",
     "@domain/entity-client-identity": "workspace:^",
     "@domain/entity-currency": "workspace:^",
     "@domain/entity-currency-crypto": "workspace:^",
     "@domain/entity-currency-fiat": "workspace:^",
     "@domain/entity-currency-token": "workspace:^",
     "@domain/entity-currency-unit": "workspace:^",
+    "@domain/entity-interest-rate": "workspace:^",
     "@features/platform-env": "workspace:^",
     "@features/platform-feature-flags": "workspace:^",
     "@ledgerhq/asset-aggregation": "workspace:^",
