+++ libs/ledger-live-common/src/dada-client/state-manager/types.ts
@@ -1,47 +1,13 @@
-import { AssetsData } from "../entities";
-
-export enum AssetsDataTags {
-  Assets = "Assets",
-}
-
-export enum AssetsAdditionalData {
-  Apy = "apy",
-  MarketTrend = "marketTrend",
-}
-
-export enum AssetCategory {
-  Stablecoins = "stablecoins",
-  Stocks = "stocks",
-}
-
-export interface GetAssetsDataParams {
-  search?: string;
-  currencyIds?: string[];
-  networkIds?: readonly string[];
-  categories?: AssetCategory[];
-  useCase?: string;
-  product: "llm" | "lld";
-  version: string;
-  isStaging?: boolean;
-  additionalData?: AssetsAdditionalData[];
-  includeTestNetworks?: boolean;
-}
-
-export interface PageParam {
-  cursor?: string;
-}
-
-export interface AssetsDataWithPagination extends AssetsData {
-  pagination: {
-    nextCursor?: string;
-  };
-}
-
-export const ONE_DAY_IN_SECONDS = 86_400;
-
-export interface GetAssetsByCategoryParams {
-  category: AssetCategory;
-  product: "llm" | "lld";
-  version: string;
-  isStaging?: boolean;
-}
+export {
+  AssetCategory,
+  AssetsAdditionalData,
+  AssetsDataTags,
+  ONE_DAY_IN_SECONDS,
+} from "@domain/api-aggregated-assets";
+export type {
+  AssetsData,
+  AssetsDataWithPagination,
+  GetAssetsByCategoryParams,
+  GetAssetsDataParams,
+  PageParam,
+} from "@domain/api-aggregated-assets";
