+++ libs/ledger-live-common/src/dada-client/types/trend.ts
@@ -1,1 +1,1 @@
-export type ApyType = "NRR" | "APY" | "APR";
+export type { ApyType } from "@domain/entity-interest-rate";
